backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pdfminer.six
python-docx
